requirements.txt
firecrawl-py
openai
tqdm
colorama
